rename actions/ama to decisionReview and constants
