+++ app/models/end_product.rb
@@ -21,7 +21,7 @@ class EndProduct
     "040HDER" => "Supplemental Claim Rating DTA"
   }.freeze
 
-  AMA_CODES = {
+  DECISION_REVIEW_CODES = {
     "030HLRR" => "Higher-Level Review Rating",
     "030HLRNR" => "Higher-Level Review Nonrating",
     "040SCR" => "Supplemental Claim Rating",
@@ -59,7 +59,7 @@ class EndProduct
     "070RMBVAGPMC" => "PMC Remand with BVA Grant"
   }.freeze
 
-  CODES = DISPATCH_CODES.merge(RAMP_CODES).merge(AMA_CODES).merge(DTA_CODES)
+  CODES = DISPATCH_CODES.merge(RAMP_CODES).merge(DECISION_REVIEW_CODES).merge(DTA_CODES)
 
   DISPATCH_MODIFIERS = %w[070 071 072 073 074 075 076 077 078 079 170 171 175 176 177 178 179 172].freeze
