+++ client/app/intake/pages/higherLevelReview/review.jsx
@@ -14,7 +14,7 @@ import {
   setClaimant,
   setPayeeCode,
   setLegacyOptInApproved
-} from '../../actions/ama';
+} from '../../actions/decisionReview';
 import { setReceiptDate } from '../../actions/intake';
 import { PAGE_PATHS, INTAKE_STATES, BOOLEAN_RADIO_OPTIONS, FORM_TYPES } from '../../constants';
 import { getIntakeStatus } from '../../selectors';
